Separate weighted/mod.rs and weignted_index.rs modules
